Don't lost continuation on multiple FlushAsync call
