+++ src/System.IO.Pipelines/src/System/IO/Pipelines/Pipe.cs
@@ -243,10 +243,13 @@ namespace System.IO.Pipelines
             // Always move the read tail to the write head
             _readTail = _writingHead;
             _readTailIndex = _writingHead.End;
+
+            long oldLength = _length;
             _length += _currentWriteLength;
 
             // Do not reset if reader is complete
             if (_pauseWriterThreshold > 0 &&
+                oldLength < _pauseWriterThreshold &&
                 _length >= _pauseWriterThreshold &&
                 !_readerCompletion.IsCompleted)
             {
